+++ RedGpuSDK/redgpu_x_print.h
@@ -115,6 +115,8 @@
   (unsigned)(x) == RED_PROCEDURE_ID_redCreateArrayTimestamp                        ? "RED_PROCEDURE_ID_redCreateArrayTimestamp"                        :\
   (unsigned)(x) == RED_PROCEDURE_ID_redDestroyArrayTimestamp                       ? "RED_PROCEDURE_ID_redDestroyArrayTimestamp"                       :\
   (unsigned)(x) == RED_PROCEDURE_ID_redArrayTimestampRead                          ? "RED_PROCEDURE_ID_redArrayTimestampRead"                          :\
+  (unsigned)(x) == RED_PROCEDURE_ID_redStructsMemoryAllocateWithInlineSamplers         ? "RED_PROCEDURE_ID_redStructsMemoryAllocateWithInlineSamplers"         :\
+  (unsigned)(x) == RED_PROCEDURE_ID_redStructsMemoryAllocateSamplersWithInlineSamplers ? "RED_PROCEDURE_ID_redStructsMemoryAllocateSamplersWithInlineSamplers" :\
   (unsigned)(x) == REDX_PROCEDURE_ID_redXMemoryCallbackBudgetChangeRegister        ? "REDX_PROCEDURE_ID_redXMemoryCallbackBudgetChangeRegister"        :\
   (unsigned)(x) == REDX_PROCEDURE_ID_redXMemoryCallbackBudgetChangeUnregister      ? "REDX_PROCEDURE_ID_redXMemoryCallbackBudgetChangeUnregister"      :\
   (unsigned)(x) == REDX_PROCEDURE_ID_redXMemoryReserve                             ? "REDX_PROCEDURE_ID_redXMemoryReserve"                             :\